+++ src/InstallerException.php
@@ -1,0 +1,6 @@
+<?php
+namespace Inet\SugarCRM;
+
+class InstallerException extends \RuntimeException
+{
+}