+++ CHANGELOG.md
@@ -4,4 +4,5 @@ Changelog
 1.X.Y
 -----
 * Add LangFileCleaner class to sort and clean SugarCRM language files and make it easier for VCS
+* Add Installer class to extract and install sugar from a zip file.